gas_tracker/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.1.0
bitarray==2.9.1
certifi==2023.11.17
charset-normalizer==3.3.2
cytoolz==0.12.2
Django==5.0
eth-abi==4.2.1
eth-account==0.10.0
eth-hash==0.5.2
eth-keyfile==0.7.0
eth-keys==0.4.0
eth-rlp==1.0.0
eth-typing==3.5.2
eth-utils==2.3.1
frozenlist==1.4.1
hexbytes==0.3.1
idna==3.6
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
lru-dict==1.2.0
multidict==6.0.4
parsimonious==0.9.0
protobuf==4.25.1
pycryptodome==3.19.1
python-dotenv==1.0.0
pyunormalize==15.1.0
referencing==0.32.0
regex==2023.12.25
requests==2.31.0
rlp==4.0.0
rpds-py==0.15.2
sqlparse==0.4.4
toolz==0.12.0
typing_extensions==4.9.0
urllib3==2.1.0
web3==6.13.0
websockets==12.0
yarl==1.9.4
